requirements.txt
# --- Core LLM / Agent ---
langchain>=1.0,<2.0
langchain-core>=1.0,<2.0
langchain-openai>=1.0,<2.0
langchain-community>=0.4,<1.0

# --- Optional provider: local models via Ollama ---
# Uncomment (or install on demand) to use local LLMs instead of OpenAI.
# langchain-ollama>=0.2

# --- Document parsing ---
pypdf>=5.0
python-docx>=1.1            # DOCX résumé support

# --- OCR (optional) — install system tesseract to use ---
# pytesseract>=0.3.10
# pdf2image>=1.17

# --- Data, output ---
pandas>=2.2
openpyxl>=3.1
pydantic>=2.7

# --- CLI / UX ---
rich>=13.9
python-dotenv>=1.0
tenacity>=9.0               # retry with exponential backoff
tiktoken>=0.8               # token counting for cost estimate
